[Fix] Fixes Issue #145 on MTG/sms-tools
`hfreqp` is initialized to a python empty list. It gets reassigned to be a Numpy array in the loop to calculate frame-wise harmonics, and thus the comparison fails to work.
Modifying the code as per this PR, makes the comparison work both for empty Python lists and empty Numpy arrays.

Signed-off-by: Anmol Mishra <anmolmishra1997@gmail.com>